requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cloudinary==1.24.0
dj-database-url==0.5.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-heroku==0.3.1
django-model-utils==4.1.1
django-ratings==0.3.7
django-star-ratings==0.9.1
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2020.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.26.2
whitenoise==5.2.0
